# Add ARM32 kernel implementation