--- a/larq_compute_engine/core/bgemm_kernels_arm32.h
+++ b/larq_compute_engine/core/bgemm_kernels_arm32.h
@@ -524,3 +524,4 @@ void BinaryKernelNeonOutOfOrder32BP4x4(BinaryKernelParams<4, 4, std::uint32_t>& 
 #undef RUY_OFFSET_DST_BASE_PTR
 
 #endif  // RUY_PLATFORM(NEON_64) && RUY_OPT_ENABLED(RUY_OPT_ASM)
+#endif  // COMPUTE_EGNINE_TFLITE_KERNELS_BGEMM_KERNELS_ARM_H_