--- a/larq_compute_engine/core/bgemm_kernels_arm32.h
+++ b/larq_compute_engine/core/bgemm_kernels_arm32.h
@@ -158,7 +158,7 @@ void BinaryKernelNeonOutOfOrder32BP4x4(
       "ldr r3, [%[params], #" RUY_STR(RUY_OFFSET_START_COL) "]\n"
       "str r3, [sp, #" RUY_STR(RUY_STACK_OFFSET_COL) "]\n"
 
-       //auto lhs_col_ptr = params.lhs_base_ptr
+      // auto lhs_col_ptr = params.lhs_base_ptr
       "ldr r1, [%[params], #" RUY_STR(RUY_OFFSET_LHS_BASE_PTR) "]\n"
       "str r1, [sp, #" RUY_STR(RUY_STACK_OFFSET_LHS_COL_PTR) "]\n"