--- a/larq_compute_engine/tflite/tests/bconv2d_test.cc
+++ b/larq_compute_engine/tflite/tests/bconv2d_test.cc
@@ -276,13 +276,9 @@ struct TestParam {
 };
 
 const auto kKernelMap = new std::map<string, register_function>({
-#if RUY_PLATFORM_ARM_32
   {"BConv2D32OPT", compute_engine::tflite::Register_BCONV_2D32_OPT},
-#elif RUY_PLATFORM_ARM_64
   {"BConv2D64OPT", compute_engine::tflite::Register_BCONV_2D64_OPT},
-#else
   {"BConv2D32REF", compute_engine::tflite::Register_BCONV_2D32_REF},
-#endif
 });
 
 class BConv2DOpTest : public ::testing::TestWithParam<TestParamTuple> {