@lgeiger 2020-07-22
Thank you for the great PR!

> I can see a couple CI/CD builds failed on github. Not sure how to run the tests offline. Please let me know if you have more details of the error or some instructions I can test the build offline in my local setup. Thanks!

It looks like linting fails on CI. We use [`clang-format`](https://clang.llvm.org/docs/ClangFormat.html) to format our C++ code so running:
```
clang-format -i larq_compute_engine/core/*.h
```
should make linting on CI happy.

[We run our ARM tests on CI](https://github.com/larq/compute-engine/blob/9ff9dad95382e6268cfc96e2a566ca77bf971253/.github/workflows/unittests.yml#L34-L58) using [`qemu`](https://www.qemu.org/). To reproduce the CI failure locally on a x86 Linux system or inside an Ubuntu docker container requires a bit of setup:
```
sudo apt-get install -y --no-install-recommends qemu-user
./configure.sh <<< $'n\n'  # or run ./configure.sh and answer with "n"
pip install numpy six
```

After which you should be able to run the ARM32 unittests in the emulator:
```
bazelisk test larq_compute_engine/tests:cc_tests_arm32_qemu --config=rpi3 --test_output=streamed
```