@honglh 2020-07-30
Thanks all for your suggestions. I will review and make a new PR