@lgeiger 2020-07-30
> I will review and make a new PR

:+1: Thanks for taking a look.
Feel free to push the updates to this branch which might be easier for you instead of creating a new PR.