@lgeiger 2020-07-30
@honglh Thanks for addressing the comments. It looks like some of the kernel tests for the optimized code are still failing on CI, could you take a look at them?
I think you should be able to reproduce the failures locally with the commands mentioned above, please let us know if you run into any issues with that.