@honglh 2020-08-02
With change in [bconv2d_test.cc](https://github.com/larq/compute-engine/pull/432/files#diff-d71ef726ccbac8c12e8216c74ed0b86b) the test execution 

```
bazel test --cache_test_results=no larq_compute_engine/tests:cc_tests_arm32_qemu --config=rpi3 --test_output=streamed
```

**passed** in my local setup.  The earlier CI failure seems caused by `Register_BCONV_2D64_OPT()` in ARM32 (`--rpi3`) mode.  In this mode `kRuyOptimized` path is selected. Then in `EvalChooseKernelType()` the `kRuyOptimized` path selects the ARM32 kernel as expected, but the `prepare()` function in `Register_BCONV_2D64_OPT()` used the function for 64-bit bitpacking bitwidth 

```
bconv2d::Prepare<bconv2d::KernelType::kRuyOptimized, 64>
``` 

This mismatch (64-bit bitpacking and 32-bit bgemm kernel) caused CI failures.

Another failure case is from `OptimizedKernel16BitOverflowTest,` which seems to be a 64-bit only test case in that the kernel `Register_BCONV_2D64_OPT` is explicitly specified. Thus this test  can is also disabled for 32-bit.